refactor: Cart 어댑터의 ViewModel 의존성 제거
