+++ app/src/main/java/woowacourse/shopping/feature/cart/cartdetail/CartItemClickHandler.kt
@@ -1,0 +1,9 @@
+package woowacourse.shopping.feature.cart.cartdetail
+
+import woowacourse.shopping.domain.model.CartItem
+
+interface CartItemClickHandler {
+    fun itemDelete(cartItem: CartItem)
+
+    fun toggleSelected(cartItem: CartItem)
+}
